+++ test/test.ts
@@ -124,6 +124,16 @@ describe("Zod v3 to v4", () => {
     });
   });
 
+  describe("Zod types", () => {
+    it("removes deleted types", async () => {
+      await runScenario("zod-types.deleted");
+    });
+
+    it("removes deleted types without importing them if prefixed", async () => {
+      await runScenario("zod-types.deleted-prefixed");
+    });
+  });
+
   it("handles import aliases", async () => {
     await runScenario("import-aliases");
   });